fix: organize request and response files
