+++ core/amber/src/main/scala/edu/uci/ics/texera/web/model/http/response/result/ResultExportResponse.scala
@@ -1,3 +1,3 @@
-package edu.uci.ics.texera.web.model.websocket.response
+package edu.uci.ics.texera.web.model.http.response.result
 
 case class ResultExportResponse(status: String, message: String)
